色々修正、controller分割
